ChatGPT API/requirements.txt
flask
gemini API
python-dotenv
